Tests shifted from instrument to commonbase.
